cache fix

+++ src/brutes/client/net/NetworkClient.java
@@ -256,10 +256,12 @@ public class NetworkClient extends Network{
     
     public static void clearCache(){
         File cacheDir = new File(NetworkClient.CACHE_DIR);
-        String[] files = cacheDir.list();
-        for (int i = 0; i < files.length; i++){
-            new File(cacheDir, files[i]).delete();
+        if(cacheDir.exists()){
+            String[] files = cacheDir.list();
+            for (int i = 0; i < files.length; i++){
+                new File(cacheDir, files[i]).delete();
+            }
+            cacheDir.delete();
         }
-        cacheDir.delete();
     }
 }
